+++ config/routes.rb
@@ -12,6 +12,8 @@ WwwSleepdataOrg::Application.routes.draw do
       get "(/a/:auth_token)/manifest", action: 'manifest', as: :manifest
       get "files((/a/:auth_token)(/m/:medium)/*path)", action: 'files', as: :files, format: false
       get "pages(/*path)", action: 'pages', as: :pages, format: false
+      get "edit_page/*path", action: 'edit_page', as: :edit_page, format: false
+      patch "update_page/*path", action: 'update_page', as: :update_page, format: false
     end
   end